requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
python-multipart
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==4.0.1